Create personal-resolv.conf
I save it here because it keeps getting overwritten by my distro on every reboot so I have created a script which will copy this file and re-write it to /etc/resolv.conf upon logging in to the system. Very annoying, but its a viable solution which works for me at the moment.